Cleanup with small refactor to retain functionality
