+++ manage/index.mjs
@@ -1,7 +1,38 @@
-import getActiveProject from '../utilities/getActiveProject.mjs'
-import renderRoles from '../utilities/renderRoles.mjs'
+// import getActiveProject from '../utilities/getActiveProject.mjs'
+// import renderRoles from '../utilities/renderRoles.mjs'
+import Project from "../api/Project.mjs"
+import TPEN from "../api/TPEN.mjs"
+import { eventDispatcher } from "../api/events.mjs"
 
-document.addEventListener("DOMContentLoaded", () => {
+eventDispatcher.on('tpen-project-loaded', loadManageInterface)
+
+async function getActiveProject() {
+    const URLParams = new URLSearchParams(window.location.search)
+    let projectID = URLParams.get("projectID")
+    let projectObj = new Project(projectID) 
+    const projectData = await projectObj.fetch() 
+    return { projectObj, projectData }
+}
+
+function renderRoles(roles) {
+    const defaultRoles = ["OWNER", "LEADER", "CONTRIBUTOR", "VIEWER"]
+
+    return roles
+        .map(role => {
+            if (role === "OWNER") {
+                return `<span class="role owner">Owner</span>`
+            } else if (role === "LEADER") {
+                return `<span class="role leader">Leader</span>`
+            } else if (defaultRoles.includes(role)) {
+                return `<span class="role default">${role.toLowerCase()}</span>`
+            } else {
+                return `<span class="role custom">${role.toLowerCase().replaceAll("_", " ")}</span>`
+            }
+        })
+        .join(", ")
+}
+
+function loadManageInterface(){
   const buttons = document.querySelectorAll("header button")
   const content = document.getElementById("content")
 
@@ -46,4 +77,6 @@ document.addEventListener("DOMContentLoaded", () => {
     })
   })
   document.getElementById("projectsHighlight").click()
-})
+}
+
+document.addEventListener("DOMContentLoaded", () => getActiveProject)
