--- projects/index.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="../css/project.css" />
-    <script src="./index.mjs" type="module"></script>
-    <script type="module" src="../components/public-login/index.js"></script>
-    <title>My Projects</title>
-  </head>
-  <body>
-  <div class="app-bg" >
-    <auth-button></auth-button>
-    <header>
-      <h3>My Projects (Hover on items to see options)</h3>
-    </header>
-    <script src="../components/projects-list/index.mjs" type="module"></script>
-    <tpen-projects-list></tpen-projects-list>
-    <div class="projects-container"  >
-      <ul class="project-list" id="projects-container">
-        <li>
-          Loading projects...
-        </li>
-      </ul>
-      <div class="last-edited">
-        <div class="image">
-          <figure>
-            <img src="../assets/images/manuscript.webp" alt="">
-          </figure>
-          <div class="caption">Medieval manuscript.webp</div>
-        </div>
-      </div>
-    </div>
-  </div>
-  </body>
-</html>
