--- utilities/getHash.mjs
@@ -1,8 +1,0 @@
-export default function getHash(agent) {
-    if (agent) {
-      return agent.split("id/").pop();
-    } else {
-      throw new Error("No agent provided");
-    }
-  }
-
