--- utilities/renderRoles.mjs
@@ -1,27 +1,0 @@
-// export default function renderRoles(roles) {
-//     if (roles.includes("OWNER")) {
-//         return "owner"
-//     } else if (roles.includes("LEADER")) {
-//         return "leader"
-//     } else {
-//         return roles.join(", ").toLowerCase().replaceAll("_", " ")
-//     }
-// }
-
-export default function renderRoles(roles) {
-    const defaultRoles = ["OWNER", "LEADER", "CONTRIBUTOR", "VIEWER"]
-
-    return roles
-        .map(role => {
-            if (role === "OWNER") {
-                return `<span class="role owner">Owner</span>`
-            } else if (role === "LEADER") {
-                return `<span class="role leader">Leader</span>`
-            } else if (defaultRoles.includes(role)) {
-                return `<span class="role default">${role.toLowerCase()}</span>`
-            } else {
-                return `<span class="role custom">${role.toLowerCase().replaceAll("_", " ")}</span>`
-            }
-        })
-        .join(", ")
-}
